backend/requirements.txt
fastapi==0.110.0
pydantic==2.6.1
uvicorn[standard]==0.23.2
starlette==0.36.3
pydantic[email]==2.6.1
pymongo==4.6.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
pytest==8.4.1
